+++ app/src/main/java/com/example/sisyphus/Model/HabitEventListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.sisyphus;
+package com.example.sisyphus.Model;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -10,9 +10,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import org.w3c.dom.Text;
+import com.example.sisyphus.R;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 
 public class HabitEventListAdapter extends ArrayAdapter<HabitEvent> {
