+++ app/src/main/java/com/example/sisyphus/Model/User.java
@@ -1,4 +1,4 @@
-package com.example.sisyphus;
+package com.example.sisyphus.Model;
 
 import java.util.Date;
