+++ app/src/main/java/com/example/sisyphus/View/EmptyMainMenu.java
@@ -1,19 +1,14 @@
-package com.example.sisyphus;
+package com.example.sisyphus.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
-import com.google.android.gms.tasks.OnSuccessListener;
+import com.example.sisyphus.R;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.io.Serializable;
 
 public class EmptyMainMenu extends AppCompatActivity {
     final String TAG = "Sample";
@@ -32,7 +27,7 @@ public class EmptyMainMenu extends AppCompatActivity {
         button_allHabitList.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(EmptyMainMenu.this,AllHabitListView.class);
+                Intent intent = new Intent(EmptyMainMenu.this, AllHabitListView.class);
                 startActivity(intent);
             }
         });
