+++ app/src/main/java/com/example/sisyphus/View/Entry.java
@@ -1,4 +1,4 @@
-package com.example.sisyphus;
+package com.example.sisyphus.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,6 +6,10 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.sisyphus.R;
+import com.example.sisyphus.View.Register;
+import com.example.sisyphus.View.SignIn;
 
 public class Entry extends AppCompatActivity {
