+++ app/src/main/java/com/example/sisyphus/View/MainActivity.java
@@ -1,23 +1,13 @@
-package com.example.sisyphus;
+package com.example.sisyphus.View;
 
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.text.Editable;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
 
 
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.util.Date;
+import com.example.sisyphus.R;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -28,7 +18,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        Intent intent = new Intent(MainActivity.this,Entry.class);
+        Intent intent = new Intent(MainActivity.this, Entry.class);
         startActivity(intent);
     }
