+++ app/src/main/res/layout/activity_add_habit.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AddHabit">
+    tools:context=".View.AddHabit">
 
     <include layout="@layout/actionbar"/>
     <com.google.android.material.textfield.TextInputLayout
