+++ app/src/main/res/layout/activity_add_habit_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddHabitEvent">
+    tools:context=".View.AddHabitEvent">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/bottom_bar2"
