+++ app/src/main/res/layout/activity_all_habit_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AllHabitListView">
+    tools:context=".View.AllHabitListView">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/top_bar"
