+++ app/src/main/res/layout/activity_edit_habit_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditHabitEventView">
+    tools:context=".View.EditHabitEventView">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/bottom_bar2"
