+++ app/src/main/res/layout/activity_empty_main_menu.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".EmptyMainMenu">
+    tools:context=".View.EmptyMainMenu">
 
 <TextView
     android:id="@+id/homepagetext"
